refactor player weapon
