+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm_Idle.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7f0fb0cde58d485aaee8ff4d1d74e3a5
+timeCreated: 1759496902
