+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm_ImpaleRecovery.cs
@@ -1,0 +1,12 @@
+public partial class PlayerWeaponFsm
+{
+    private void ImpaleRecoveryConfigure()
+    {
+        Machine.Configure(PlayerWeaponFsmState.ImpaleRecovery)
+            .Permit(FsmTrigger.Timeout, PlayerWeaponFsmState.Idle)
+            .OnEntry(_ =>
+            {
+                // transform.DOShakePosition(0.5f, 0.3f);
+            });
+    }
+}
