+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm_ImpaleStartup.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ff40defc5fbc44659833727be3fab4db
+timeCreated: 1759496892
