+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm_ImpaleStuck.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: caa9d8b138654bae87453cfa0bb9ad99
+timeCreated: 1759496849
