+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm__common.cs
@@ -1,0 +1,25 @@
+using System;
+using Cinemachine;
+using UnityEngine;
+
+public partial class PlayerWeaponFsm
+{
+    private Transform _subTransform;
+    private Vector3 _subTransformBaseLocalPosition;
+    private Vector3 _impaleActiveTargetPosition;
+    private CinemachineImpulseSource _impulseSource;
+    public static PlayerWeaponFsm Singleton;
+    private const float IdleOrbitRadius = 3f;
+    private const float IdleOrbitHeight = 3.5f;
+    private const float IdlePositionLerpStrength = 9f;
+    private const float IdleRotationLerpStrength = 5f;
+    private const float ImpaleStartupOrbitRadius = 1f;
+    private const float ImpaleStartupOrbitHeight = 4.5f;
+    private const float ImpaleStartupPositionLerpStrength = 27.5f;
+    private const float ImpaleStartupRotationLerpStrength = 15f;
+    private const float ImpaleStartupPullbackSpeed = 6.5f;
+    private const float ImpaleActivePositionLerpStrength = 45f;
+    private const float ImpaleActiveMaxDistance = 10f;
+    private const float ImpaleStuckRecoveryPullbackSpeed = 14f;
+    public static event Action<Vector3, bool> OnPlayerWeaponPositionUpdated;
+}
