+++ Assets/Code/Fsm/PlayerWeaponFsm/PlayerWeaponFsm__setup.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a8dd399970494085acfadcbb2cbc2b12
+timeCreated: 1759496563
