+++ Assets/Graphics/Shader/Triplanar_Checkerboard 1.mat
@@ -59,12 +59,12 @@ Material:
         m_Offset: {x: 0, y: 0}
     m_Ints: []
     m_Floats:
-    - _Emission: 1
+    - _Emission: 0
     - _Invert: 0.45
     - _QueueControl: 0
     - _QueueOffset: 0
     - _Scale: 0.3
     m_Colors:
-    - _Color: {r: 0.46226418, g: 0.46226418, b: 0.46226418, a: 0}
+    - _Color: {r: 1, g: 1, b: 1, a: 0}
     - _Top_Color: {r: 1, g: 1, b: 1, a: 1}
   m_BuildTextureStacks: []
