+++ Assets/Graphics/Shader/Triplanar_Checkerboard Variant.mat
@@ -26,9 +26,9 @@ Material:
     m_TexEnvs: []
     m_Ints: []
     m_Floats:
-    - _Emission: 1
+    - _Emission: 0
     m_Colors:
-    - _Color: {r: 0.4528302, g: 0.35637736, b: 0.38621396, a: 0}
+    - _Color: {r: 0.6415094, g: 0.6415094, b: 0.6415094, a: 0}
   m_BuildTextureStacks: []
 --- !u!114 &2407142867973555347
 MonoBehaviour:
