+++ Assets/Scenes/Test.unity
@@ -14,18 +14,18 @@ OcclusionCullingSettings:
 RenderSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 9
-  m_Fog: 0
-  m_FogColor: {r: 1, g: 1, b: 1, a: 1}
-  m_FogMode: 3
-  m_FogDensity: 0.005
-  m_LinearFogStart: 0
-  m_LinearFogEnd: 300
-  m_AmbientSkyColor: {r: 0, g: 0.5513765, b: 1.4142135, a: 1}
+  m_Fog: 1
+  m_FogColor: {r: 0, g: 0.9254899, b: 1, a: 1}
+  m_FogMode: 1
+  m_FogDensity: 0.0325
+  m_LinearFogStart: 40
+  m_LinearFogEnd: 1400
+  m_AmbientSkyColor: {r: 0.79341686, g: 0.79341686, b: 0.79341686, a: 1}
   m_AmbientEquatorColor: {r: 0.114, g: 0.125, b: 0.133, a: 1}
   m_AmbientGroundColor: {r: 0.047, g: 0.043, b: 0.035, a: 1}
   m_AmbientIntensity: 0
   m_AmbientMode: 3
-  m_SubtractiveShadowColor: {r: 0.41709685, g: 0.5420195, b: 0.8584906, a: 1}
+  m_SubtractiveShadowColor: {r: 1, g: 1, b: 1, a: 1}
   m_SkyboxMaterial: {fileID: 10304, guid: 0000000000000000f000000000000000, type: 0}
   m_HaloStrength: 0.5
   m_FlareStrength: 1
@@ -1618,7 +1618,7 @@ PrefabInstance:
     - target: {fileID: 7543281101885970428, guid: 51b04cec7607344fb8a4c598bd9936d4,
         type: 3}
       propertyPath: m_LookaheadTime
-      value: 0.244
+      value: 0.3
       objectReference: {fileID: 0}
     - target: {fileID: 7543281101885970428, guid: 51b04cec7607344fb8a4c598bd9936d4,
         type: 3}
@@ -5792,7 +5792,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1632466210}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 172515602e62fb746b5d573b38a5fe58, type: 3}
   m_Name:
